// apps/collector/src/generated/twitch-api.d.ts
export interface paths {
// ...
    "/games": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Gets information about specified games.
         * @description Gets information about specified categories or games.
         *
         *     You may get up to 100 categories or games by specifying their ID or name. You may specify all IDs, all names, or a combination of IDs and names. If you specify a combination of IDs and names, the total number of IDs and names must not exceed 100.
         *
         *     __Authorization:__
         *
         *     Requires an [app access token](https://dev.twitch.tv/docs/authentication#app-access-tokens) or [user access token](https://dev.twitch.tv/docs/authentication#user-access-tokens).
         */
        get: operations["get-games"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/streams": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Gets a list of all streams.
         * @description Gets a list of all streams. The list is in descending order by the number of viewers watching the stream. Because viewers come and go during a stream, it’s possible to find duplicate or missing streams in the list as you page through the results.
         *
         *     __Authorization:__
         *
         *     Requires an [app access token](https://dev.twitch.tv/docs/authentication#app-access-tokens) or [user access token](https://dev.twitch.tv/docs/authentication#user-access-tokens).
         */
        get: operations["get-streams"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/users": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Gets information about one or more users.
         * @description Gets information about one or more users.
         *
         *     You may look up users using their user ID, login name, or both but the sum total of the number of users you may look up is 100\. For example, you may specify 50 IDs and 50 names or 100 IDs or names, but you cannot specify 100 IDs and 100 names.
         *
         *     If you don’t specify IDs or login names, the request returns information about the user in the access token if you specify a user access token.
         *
         *     To include the user’s verified email address in the response, you must use a user access token that includes the **user:read:email** scope.
         *
         *     __Authorization:__
         *
         *     Requires an [app access token](https://dev.twitch.tv/docs/authentication#app-access-tokens) or [user access token](https://dev.twitch.tv/docs/authentication#user-access-tokens).
         */
        get: operations["get-users"];
        /**
         * Updates the user’s information.
         * @description Updates the specified user’s information. The user ID in the OAuth token identifies the user whose information you want to update.
         *
         *     To include the user’s verified email address in the response, the user access token must also include the **user:read:email** scope.
         *
         *     __Authorization:__
         *
         *     Requires a [user access token](https://dev.twitch.tv/docs/authentication#user-access-tokens) that includes the **user:edit** scope.
         */
        put: operations["update-user"];
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        Stream: {
            /** @description An ID that identifies the stream. You can use this ID later to look up the video on demand (VOD). */
            id: string;
            /** @description The ID of the user that’s broadcasting the stream. */
            user_id: string;
            /** @description The user’s login name. */
            user_login: string;
            /** @description The user’s display name. */
            user_name: string;
            /** @description The ID of the category or game being played. */
            game_id: string;
            /** @description The ID of the category or game being played. */
            game_name: string;
            /**
             * @description The type of stream. Possible values are:
             *
             *     * live
             *
             *     If an error occurs, this field is set to an empty string.
             * @enum {string}
             */
            type: "live";
            /** @description The stream’s title. Is an empty string if not set. */
            title: string;
            /**
             * Format: int32
             * @description The number of users watching the stream.
             */
            viewer_count: number;
            /**
             * Format: date-time
             * @description The UTC date and time (in RFC3339 format) of when the broadcast began.
             */
            started_at: string;
            /** @description The language that the stream uses. This is an ISO 639-1 two-letter language code or _other_ if the stream uses a language not in the list of [supported stream languages](https://help.twitch.tv/s/article/languages-on-twitch#streamlang). */
            language: string;
            /** @description A URL to an image of a frame from the last 5 minutes of the stream. Replace the width and height placeholders in the URL (`{width}x{height}`) with the size of the image you want, in pixels. */
            thumbnail_url: string;
            /**
             * @deprecated
             * @description **IMPORTANT** As of February 28, 2023, this field is deprecated and returns only an empty array. If you use this field, please update your code to use the `tags` field.
             *
             *     The list of tags that apply to the stream. The list contains IDs only when the channel is steaming live. For a list of possible tags, see [List of All Tags](https://www.twitch.tv/directory/all/tags). The list doesn’t include Category Tags.
             */
            tag_ids: string[];
            /** @description The tags applied to the stream. */
            tags: string[];
            /**
             * @description **IMPORTANT** This field is deprecated and returns only `false`.
             *
             *     A Boolean value that indicates whether the stream is meant for mature audiences.
             */
            is_mature: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    "start-commercial": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
// ...
    };
// ...
}